focusmode is 16bit per default, not 8bit. this might break big endian https://github.com/gphoto/gphoto2/issues/268

+++ camlibs/ptp2/library.c
@@ -4235,7 +4235,7 @@ camera_sony_capture (Camera *camera, CameraCaptureType type, CameraFilePath *pat
 
 	/* Check if we are in manual focus to skip the wait for focus */
 	C_PTP (ptp_generic_getdevicepropdesc (params, PTP_DPC_FocusMode, &dpd));
-	if (dpd.CurrentValue.u8 == 2) {
+	if (dpd.CurrentValue.u16 == 2) {
 
 		/* Now hold down the shutter button for a bit. We probably need to hold it as long as it takes to
 		* get focus, indicated by the 0xD213 property. But hold it for at most 1 second.
@@ -5439,7 +5439,7 @@ camera_trigger_capture (Camera *camera, GPContext *context)
 
 		/* Wait for focus only in automatic focus mode */
 		C_PTP (ptp_generic_getdevicepropdesc (params, PTP_DPC_FocusMode, &dpd));
-		if (dpd.CurrentValue.u8 == 2) {
+		if (dpd.CurrentValue.u16 == 2) {
 
 			/* Now hold down the shutter button for a bit. We probably need to hold it as long as it takes to
 			* get focus, indicated by the 0xD213 property. But hold it for at most 1 second.
